Split up Decoupled.scala into multiple files, NFC
Break up Decoupled.scala into four files, each associated with its
internal contents.  This makes it easier for developers to find Queue,
Irrevocable, and ReadyValidIO.
