+++ src/main/scala/chisel3/util/ReadyValidIO.scala
@@ -1,0 +1,83 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package chisel3.util
+
+import chisel3._
+import chisel3.experimental.{requireIsChiselType, Direction}
+import chisel3.reflect.DataMirror
+import chisel3.util.simpleClassName
+
+/** An I/O Bundle containing 'valid' and 'ready' signals that handshake
+  * the transfer of data stored in the 'bits' subfield.
+  * The base protocol implied by the directionality is that
+  * the producer uses the interface as-is (outputs bits)
+  * while the consumer uses the flipped interface (inputs bits).
+  * The actual semantics of ready/valid are enforced via the use of concrete subclasses.
+  * @param gen the type of data to be wrapped in Ready/Valid
+  * @groupdesc Signals The actual hardware fields of the Bundle
+  */
+abstract class ReadyValidIO[+T <: Data](gen: T) extends Bundle {
+
+  /** Indicates that the consumer is ready to accept the data this cycle
+    * @group Signals
+    */
+  val ready = Input(Bool())
+
+  /** Indicates that the producer has put valid data in 'bits'
+    * @group Signals
+    */
+  val valid = Output(Bool())
+
+  /** The data to be transferred when ready and valid are asserted at the same cycle
+    * @group Signals
+    */
+  val bits = Output(gen)
+
+  /** A stable typeName for this `ReadyValidIO` and any of its implementations
+    * using the supplied `Data` generator's `typeName`
+    */
+  override def typeName = s"${simpleClassName(this.getClass)}_${gen.typeName}"
+}
+
+object ReadyValidIO {
+
+  implicit class AddMethodsToReadyValid[T <: Data](target: ReadyValidIO[T]) {
+
+    /** Indicates if IO is both ready and valid
+      */
+    def fire: Bool = target.ready && target.valid
+
+    /** Push dat onto the output bits of this interface to let the consumer know it has happened.
+      * @param dat the values to assign to bits.
+      * @return    dat.
+      */
+    def enq(dat: T): T = {
+      target.valid := true.B
+      target.bits := dat
+      dat
+    }
+
+    /** Indicate no enqueue occurs. Valid is set to false, and bits are
+      * connected to an uninitialized wire.
+      */
+    def noenq(): Unit = {
+      target.valid := false.B
+      target.bits := DontCare
+    }
+
+    /** Assert ready on this port and return the associated data bits.
+      * This is typically used when valid has been asserted by the producer side.
+      * @return The data bits.
+      */
+    def deq(): T = {
+      target.ready := true.B
+      target.bits
+    }
+
+    /** Indicate no dequeue occurs. Ready is set to false.
+      */
+    def nodeq(): Unit = {
+      target.ready := false.B
+    }
+  }
+}
